requirements.txt
anytree==2.8.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
iniconfig==1.1.1
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
virtualenv==20.2.2
wrapt==1.12.1
anytree==2.8.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
idna==2.10
importlib-metadata==2.1.1
iniconfig==1.1.1
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
zipp==3.4.0
anytree==2.8.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
distlib==0.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
idna==2.10
importlib-metadata==2.1.1
iniconfig==1.1.1
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.20.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
zipp==3.4.0
